Loaded Postgres to BQ
