+++ docker-compose.yaml
@@ -15,12 +15,19 @@ x-airflow-common:
     AIRFLOW__CORE_TEST: 'true'
     AIRFLOW__CORE__TEST_CONNECTION: 'enabled'
     AIRFLOW__SCHEDULER__ENABLE_HEALTH_CHECK: 'true'
+    GOOGLE_APPLICATION_CREDENTIALS: /opt/airflow/config/altschool-capstones-e2f54a131769.json
+      # Mount the key file
     # _PIP_ADDITIONAL_REQUIREMENTS: ${_PIP_ADDITIONAL_REQUIREMENTS:- }
   volumes:
     - ${AIRFLOW_PROJ_DIR:-./airflow}/dags:/opt/airflow/dags
     - ${AIRFLOW_PROJ_DIR:-./airflow}/logs:/opt/airflow/logs
     - ${AIRFLOW_PROJ_DIR:-./airflow}/config:/opt/airflow/config
     - ${AIRFLOW_PROJ_DIR:-./airflow}/plugins:/opt/airflow/plugins
+    - ${AIRFLOW_PROJ_DIR:-./airflow}/config/altschool-capstones-e2f54a131769.json:/opt/airflow/config/altschool-capstones-e2f54a131769.json
+    - ${AIRFLOW_PROJ_DIR:-./airflow}/dags/package/schema.json:/opt/airflow/dags/package/schema.json
+    - ${AIRFLOW_PROJ_DIR:-./airflow}/config/profiles.yml:/opt/airflow/.dbt/profiles.yml
+    - ${AIRFLOW_PROJ_DIR:-./airflow}/dbt:/opt/airflow/dbt
+
   user: "${AIRFLOW_UID:-50000}:0"
   depends_on:
     &airflow-common-depends-on
