+++ requirements.txt
@@ -5,3 +5,5 @@ python-dotenv==1.0.1
 psycopg2-binary==2.9.9
 typing_extensions==4.12.1
 python-dotenv==1.0.1
+dbt-core==1.8.5
+dbt-bigquery==1.8.2
